fixed test_cli_environment.py
